.bashrc deb emacs

+++ .bashrc
@@ -233,7 +233,11 @@ if [ "$(uname)" == "Darwin" ]; then
     install emacs "emacs --with-cocoa --with-dbus --with-imagemagick@6 --with-librsvg --with-mailutils --with-modules"
     [ ! -L "${HOME}/Applications/Emacs.app" ] && ln -s "${HOMEBREW}/opt/emacs/Emacs.app" "${HOME}/Applications/"
 elif [ "$(uname)" == "Linux" ]; then
-    install emacs emacs
+    if [ -f /etc/debian_version ]; then
+        install emacs25 emacs25
+    else
+        install emacs emacs
+    fi
 fi
 
 # offlineimap + mu
